Separate actions index.js
