+++ client/src/reducers/userReducer.js
@@ -1,5 +1,14 @@
-import { GET_CURR_USER_SUCCESS, GET_CURR_USER_FAILURE, USER_START_AUTHORIZING, 
-            USER_LOGGED_IN, USER_SIGNED_UP, FAVORITE_TRAIL, FETCH_FAVORITE_TRAILS, FETCH_COMPLETED_TRAILS, USER_START_FETCHING } from '../actions/types';
+import { 
+    GET_CURR_USER_SUCCESS, 
+    GET_CURR_USER_FAILURE, 
+    USER_START_AUTHORIZING, 
+    USER_LOGGED_IN, 
+    USER_SIGNED_UP, 
+    FAVORITE_TRAIL, 
+    FETCH_FAVORITE_TRAILS, 
+    FETCH_COMPLETED_TRAILS, 
+    USER_START_FETCHING 
+} from '../actions/types';
 
 const INITIAL_STATE = null;
